service/alerts-populator/requirements.txt
asyncio-nats-client~=0.11.5
aiohttp==3.13.0
httpx~=0.28.1
pytest~=8.3.5
pytest-cov~=6.1.1
pytest-asyncio~=0.26.0
pylint~=3.3.7
pyyaml~=6.0.3
